pyproject.toml

[project]
name = "discord-voice-leveler"
version = "0.1.0"
description = "Discord Voice Level Bot"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
